split form utils function into file
